rename internationalization.py to locale.py for brevity and common meaning.
