+++ vdu_controls/vdu_controls/lux_auto.py
@@ -15,7 +15,7 @@ from vdu_controls.constants import MsgDestination
 from vdu_controls.ddcutil_abstract import BRIGHTNESS_VCP_CODE
 
 from vdu_controls.ddcutil_aggregator import VduStableId
-from vdu_controls.internationalization import tr
+from vdu_controls.locale import tr
 import vdu_controls.logging as log
 from vdu_controls.lux_ambient_slider import LuxZone, LuxAmbientSlider
 from vdu_controls.lux_config import LuxConfig, LuxPoint
